Added last modified date to the database and added the reading of the date when the audio file is opened or when tags are saved

+++ app/src/main/cpp/tagger/files/Mp3File.h
@@ -85,7 +85,7 @@ public:
      * -2 if an output stream couldn't be opened
      * 0 if successful
      */
-    int saveNewTag(Tag *newTag) override;
+    long saveNewTag(Tag *newTag) override;
 
     Tag *getTag() override;
